pyproject.toml
[project]
name = "magelab"
version = "0.1.0"
description = "Multi-agent Lab: An orchestration and experimentation framework for multi-agent LLM systems"
readme = "README.md"
license = "MIT"
authors = [
    { name = "Siddarth Srinivasan" }
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
requires-python = ">=3.12.7"
dependencies = [
    "aiohttp>=3.13.5",
    "anthropic>=0.91.0",
    "claude-agent-sdk>=0.1.56",
    "mcp>=1.27.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.3",
]

[project.scripts]
magelab = "magelab.__main__:main"

[build-system]
requires = ["uv_build>=0.11.3,<0.12"]
build-backend = "uv_build"

[tool.ruff]
line-length = 120
# extend-include = ["*.ipynb"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.pylint]
max-args = 8

[tool.ruff.format]
docstring-code-format = true
exclude = []

[dependency-groups]
dev = [
    "ipykernel>=7.2.0",
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.1.0",
]

[tool.pytest.ini_options]
    addopts = [
        "--cov=src",
        "--cov-report=term-missing",
    ]
    testpaths = ["tests"]
    asyncio_default_fixture_loop_scope = "function"
